requirements.txt
azure-ai-projects
azure-identity
flask
python-dotenv
gunicorn